Fixed message formatting.

+++ src/WorldStorage/SchematicFileSerializer.cpp
@@ -234,7 +234,7 @@ bool cSchematicFileSerializer::LoadFromSchematicNBT(cBlockArea & a_BlockArea, cP
 	if (a_NBT.GetDataLength(TBlockTypes) < NumBytes)
 	{
 		LOG("BlockTypes truncated in the schematic file (exp %d, got %d bytes). Loading partial.",
-			NumBytes, a_NBT.GetDataLength(TBlockTypes)
+			(int)NumBytes, (int)a_NBT.GetDataLength(TBlockTypes)
 		);
 		NumBytes = a_NBT.GetDataLength(TBlockTypes);
 	}
@@ -246,7 +246,7 @@ bool cSchematicFileSerializer::LoadFromSchematicNBT(cBlockArea & a_BlockArea, cP
 		if (a_NBT.GetDataLength(TBlockMetas) < NumBytes)
 		{
 			LOG("BlockMetas truncated in the schematic file (exp %d, got %d bytes). Loading partial.",
-				NumBytes, a_NBT.GetDataLength(TBlockMetas)
+				(int)NumBytes, (int)a_NBT.GetDataLength(TBlockMetas)
 			);
 			NumBytes = a_NBT.GetDataLength(TBlockMetas);
 		}
